pyproject.toml
[tool.poetry]
name = "projectjobsearch"
version = "0.1.0"
description = ""
authors = ["IvanPro91 <ivan-don4encko@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-cov = "^6.1.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20250328"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
types-psycopg2 = "^2.9.21.20250516"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
